Move get_expected_instance_count_for_namespace to long_running_service_tools
