+++ paasta_tools/marathon_tools.py
@@ -32,7 +32,6 @@ from typing import Optional
 from typing import Sequence
 from typing import Set
 from typing import Tuple
-from typing import Type
 from typing import TypeVar
 
 import pytz
@@ -58,7 +57,6 @@ from paasta_tools.mesos_tools import filter_mesos_slaves_by_blacklist
 from paasta_tools.mesos_tools import get_mesos_network_for_net
 from paasta_tools.mesos_tools import get_mesos_slaves_grouped_by_attribute
 from paasta_tools.mesos_tools import mesos_services_running_here
-from paasta_tools.paasta_service_config_loader import PaastaServiceConfigLoader
 from paasta_tools.secret_tools import get_secret_hashes
 from paasta_tools.utils import BranchDictV2
 from paasta_tools.utils import compose_job_id
@@ -1239,37 +1237,6 @@ def create_complete_config(
         cluster=load_system_paasta_config().get_cluster(),
         soa_dir=soa_dir,
     ).format_marathon_app_dict()
-
-
-def get_expected_instance_count_for_namespace(
-    service: str,
-    namespace: str,
-    cluster: str = None,
-    instance_type_class: Type[LongRunningServiceConfig] = MarathonServiceConfig,
-    soa_dir: str = DEFAULT_SOA_DIR,
-) -> int:
-    """Get the number of expected instances for a namespace, based on the number
-    of instances set to run on that namespace as specified in Marathon service
-    configuration files.
-
-    :param service: The service's name
-    :param namespace: The namespace for that service to check
-    instance_type_class: The type of the instance, options are MarathonServiceConfig and KubernetesDeploymentConfig,
-    :param soa_dir: The SOA configuration directory to read from
-    :returns: An integer value of the # of expected instances for the namespace"""
-    total_expected = 0
-    if not cluster:
-        cluster = load_system_paasta_config().get_cluster()
-
-    pscl = PaastaServiceConfigLoader(
-        service=service, soa_dir=soa_dir, load_deployments=False
-    )
-    for job_config in pscl.instance_configs(
-        cluster=cluster, instance_type_class=instance_type_class
-    ):
-        if f"{service}.{namespace}" in job_config.get_registrations():
-            total_expected += job_config.get_instances()
-    return total_expected
 
 
 def get_matching_appids(
