+++ tests/test_long_running_service_tools.py
@@ -425,3 +425,45 @@ def test_get_all_namespaces():
         get_namespaces_patch.assert_any_call("rid1", soa_dir)
         get_namespaces_patch.assert_any_call("rid2", soa_dir)
         assert get_namespaces_patch.call_count == 2
+
+
+def test_get_expected_instance_count_for_namespace():
+    service = "red"
+    namespace = "rojo"
+    soa_dir = "que_esta"
+    fake_job_configs = [
+        long_running_service_tools.LongRunningServiceConfig(
+            service=service,
+            cluster="fake_cluster",
+            instance="blue",
+            config_dict={"nerve_ns": "rojo", "instances": 11},
+            branch_dict=None,
+        ),
+        long_running_service_tools.LongRunningServiceConfig(
+            service=service,
+            cluster="fake_cluster",
+            instance="green",
+            config_dict={"nerve_ns": "amarillo"},
+            branch_dict=None,
+        ),
+    ]
+
+    with mock.patch(
+        "paasta_tools.long_running_service_tools.PaastaServiceConfigLoader",
+        autospec=True,
+        return_value=mock.Mock(
+            instance_configs=mock.Mock(return_value=fake_job_configs)
+        ),
+    ) as fake_pscl:
+        actual = long_running_service_tools.get_expected_instance_count_for_namespace(
+            service,
+            namespace,
+            cluster="fake_cluster",
+            soa_dir=soa_dir,
+            instance_type_class=long_running_service_tools.LongRunningServiceConfig,
+        )
+        assert actual == 11
+        fake_pscl.return_value.instance_configs.assert_called_once_with(
+            cluster="fake_cluster",
+            instance_type_class=long_running_service_tools.LongRunningServiceConfig,
+        )
